working fine

+++ include/sys/zio.h
@@ -453,7 +453,7 @@ struct zio {
 	/* Taskq dispatching state */
 	taskq_ent_t	io_tqent;
 
-    int rot;
+    int8_t rot;
     dnode_t *io_dn;
     const char *filp;
     //struct file *filp;
